+++ server/routes/index.js
@@ -1,6 +1,7 @@
 const route = require('express').Router()
 const todoRoute = require('./todoRoute.js')
 const userRoute = require('./userRoute.js')
+const myTodoRoute = require('./myTodoRoute')
 
 
 route.get('/', (req, res) => {
@@ -9,6 +10,7 @@ route.get('/', (req, res) => {
 
 
 route.use('/todos', todoRoute)
+route.use('/mytodos', myTodoRoute)
 route.use('/user', userRoute)
 
 module.exports = route
